requirements.txt
colorama==0.3.7
nose==1.3.7
PyYAML==3.11
requests==2.10.0
Requires==0.0.3
-e git+https://github.com/tedivm/python-screeps.git@f5b44aaaf59640ee133adaf1c62548720ab1dfe0#egg=screepsapi
six==1.10.0
urwid==1.3.1
websocket-client==0.37.0